ath10k: Fix CT ratemask handling for 4x4 NICs.
Several bugs:  First, have to extend each of the peer-assoc-complete
commands individually since otherwise we are out of sync with the
various firmware versions.

Second, don't hard-code rate-table as being 3x3.  Instead, use
the NSS reported by the firmware to know if 3x3 vs 4x4.

Third, VHT mcs entries are 16-bits, but only 10 of them
are used.

NOTE:  A good bit of this was backported to earlier patch,
but the backport missed some of the logic in this patch.
Move this back in the tree and combine with earlier
rate-set patch.

NOTE2:  Actually trying to enable the larger size for wave-2
breaks performance.  Must be more bugs or mis-merge somewhere.
Will comment that out so bisect of basic tput works OK, will
fix root cause later.

Signed-off-by: Ben Greear <greearb@candelatech.com>