mtk: mac80211: Add STA site link add support
Trigger deauth & reconnect when all the valid links of the STA receive
the information of link addition.
(max simultaneous link num in ML IE > the simultaneous link num stored
after association success)
For example, when the MLD AP completes the CAC or adds an additional link
via the link reconfiguration command, SQC/autotest might assume that the
MLD STA to react to it rather than ignoring it.

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>

Do not check ifmgd->reporting_add_links when the connection is not MLO.

This fixes the problem that an MLD STA cannot associate with an MLD AP.

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>